HealthAI_Project/requirements.txt
streamlit
google-generativeai
opencv-python-headless
mediapipe==0.10.14
pandas
numpy